requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.5
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
async-timeout==4.0.2
# backports.zoneinfo==0.2.1
django-cors-headers==4.2.0
django-crontab==0.7.1
django-model-utils==4.3.1
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
redis==4.6.0
typing-extensions==4.7.1